chore(admin-canary): refactor Tags page to use page component
