+++ packages/bonde-admin-canary/src/scenes/Logged/Root.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { AuthProvider, Route } from 'services/auth'
 import { FullScreenLoadable, LoadingFullScreen } from 'components/Loadable'
-import { Page as TagsPage } from './scenes/Tags'
+import TagsPage from './scenes/Tags'
 
 const HomePage = FullScreenLoadable({
   loader: () => import('./scenes/Home/Page')
