+++ packages/bonde-admin-canary/src/scenes/Logged/scenes/Tags/CreateUserTagsForm/UserTagsField/UserTagsField.js
@@ -2,14 +2,17 @@ import React from 'react'
 import { MultipleChoiceField, Tag } from 'bonde-styleguide'
 import { Query } from 'react-apollo'
 import { LoadingFullScreen } from 'components/Loadable'
-import TAGS from './tags.graphql'
+import UserTags from './userTags.graphql'
 
-
-const TagsField = (props) => (
-  <Query query={TAGS} variables={{ tagType: 'user' }}>
+export default ({ t, ...props }) => (
+  <Query query={UserTags}>
     {({ loading, error, data }) => {
 
-      if (loading) return <LoadingFullScreen message='Carregando tags.' />
+      if (loading) {
+        return (
+          <LoadingFullScreen message={t('loading.userTags')} />
+        )
+      }
 
       if (error) console.log('error', error)
       
@@ -28,5 +31,3 @@ const TagsField = (props) => (
     }}
   </Query>
 )
-
-export default TagsField
