+++ packages/bonde-admin-canary/src/scenes/Logged/scenes/Tags/CreateUserTagsForm/UserTagsField/index.js
@@ -1,0 +1,4 @@
+import { translate } from 'services/i18n'
+import UserTagsField from './UserTagsField'
+
+export default translate('tags')(UserTagsField)
