+++ packages/bonde-admin-canary/src/scenes/Logged/scenes/Tags/CreateUserTagsForm/UserTagsField/userTags.graphql.js
@@ -1,8 +1,8 @@
 import gql from 'graphql-tag'
 
 export default gql`
-  query tags ($tagType: String!) {
-    allTags (condition: { tagType: $tagType }) {
+  query UserTags {
+    allTags (condition: { tagType: "user" }) {
       nodes {
 	name,
 	label
