+++ packages/bonde-admin-canary/src/scenes/Logged/scenes/Tags/CreateUserTagsForm/index.js
@@ -1,0 +1,5 @@
+import { auth } from 'services/auth'
+import { translate } from 'services/i18n'
+import CreateUserTagsForm from './CreateUserTagsForm'
+
+export default translate('tags')(auth()(CreateUserTagsForm))
