--- packages/bonde-admin-canary/src/scenes/Logged/scenes/Tags/components/index.js
@@ -1,1 +1,0 @@
-export { default as TagsField } from './TagsField'
