+++ packages/bonde-admin-canary/src/scenes/Logged/scenes/Tags/index.js
@@ -1,1 +1,5 @@
-export { default as Page } from './Page'
+import { auth } from 'services/auth'
+import { translate } from 'services/i18n'
+import Tags from './Tags'
+
+export default translate('tags')(auth()(Tags))
